refactor: turn dash app into a package
