+++ Pipfile
@@ -24,6 +24,7 @@ xgboost = "==1.0.2"
 Flask = "==1.1.1"
 python-dotenv = "*"
 dash-bootstrap-components = "*"
+pyarrow = "*"
 
 [requires]
 python_version = "3.7"
