--- btc_dash/dash_layouts/insights.py
@@ -1,20 +1,0 @@
-import dash_bootstrap_components as dbc
-import dash_core_components as dcc
-
-
-column1 = dbc.Col(
-    [
-        dcc.Markdown(
-            """
-        
-            ## Insights
-            """
-        ),
-    ],
-    md=4,
-)
-
-
-column2 = dbc.Col([])
-
-layout = dbc.Row([column1, column2])
