+++ btc_dash/routes/___init__.py
@@ -1,0 +1,2 @@
+from .sitemap_route import sitemap
+from .robots_route import robots
