+++ btc_dash/templates/sitemap.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<urlset
+      xmlns="http://www.sitemaps.org/schemas/sitemap/0.9"
+      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+      xsi:schemaLocation="http://www.sitemaps.org/schemas/sitemap/0.9
+            http://www.sitemaps.org/schemas/sitemap/0.9/sitemap.xsd">
+
+{% for url in static_urls %}
+<url>
+  <loc>{{ url["loc"] }}</loc>
+</url>
+{% endfor %}
+
+<!-- {% for url in dynamic_urls %}
+<url>
+  <loc>{{ url["loc"] }}</loc>
+  <lastmod>{{ url["lastmod"] }}</lastmod>
+</url>
+{% endfor %} -->
+
+</urlset>
